requirements.txt
google-adk
pydantic
litellm
google-generativeai
python-dotenv
pydantic
google-cloud-aiplatform[adk,agent_engines]